fix(tests): fix tests for init protocol config
